authentication/requirements.txt
Flask
Werkzeug
PyJWT
flask-cors
